swndpp/requirements/base.txt
Django==1.11.3
django-model-utils==3.0.0
djangorestframework==3.6.3
olefile==0.44
Pillow==4.2.0
pkg-resources==0.0.0
psycopg2==2.7.1
pytz==2017.2
Unipath==1.1
